Execute the utility immediately after registering file watch events
@kevincox observed that we frequently switch over to our editor to save a file
just to see tests run for the first time. This is definitely the case, so it
seems worth breaking compatibility and flip to a new default.

New postpone option (-p) allows `entr` to retain the old behavior.

Another reason for this change is that the restart option already behaved this
way.