Clarify the purpose and use of the WSL / Docker for Mac inotify workaround
I found entr yesterday in the course of adding hot reloading to a Docker
development image but had trouble getting it to reload consistently.
I found this section of the README and tried the workaround, but it
didn't seem to change the behavior, so I wasn't sure if I had correctly
enabled the workaround.

entr 3.9, which I had installed from apt-get, did not yet have the
workaround feature. Instead I had to build from source. I also didn't
know to expect the warning entr prints when the workaround is enabled,
which would have given me a way to know whether my configuration was
right.

I hope these hints reduce the barrier to entry for newcomers like me.