README: fix examples of using find(1)
In commit eadd72f11eea, the examples of using ls(1) were changed to use
find(1).  Even though the examples might seam to work, the correct way
to search for file names is to use the `-name' command and for files in
general, the `type f' command.

Typically, we should change as well the very first example, which
returns *both* source files *and* directories.  The comment suggests we
are interested only in source files.

In GNU find(1) we could omit the search directory, by default is current
working directory, but this is not portable in POSIX or BSD find's.