Sync the README with the man page examples
I'm assuming the use of `-s` from the man page is newer and preferable
to the README's `sh -c "..."` formulation.

Conversely, I retained the use of `head -n 20` from the README, rather
than the man page's `sed 20q`, figuring that the operation of `head` is
likely to be familiar to more readers.

(This nitpicking sync is a precursor to a follow-up submission which
will propose adding a tiny amount of explanatory verbiage to this
documentation.)